Refactor Column
